fix: expected getBy query errors being logged
When getBy queries throw the resulting errors are not caught in the
executeAsync call, WebdriverIO thinks this error is not expected and so
logs it as an error. This is a problem for tests that expect an error to
be produced.

Catch errors thrown by getBy queries and pass their messages back like
we do for findBy promise rejections.

Change the logLevel in wdio.conf.js to 'warn' to see these logs in
future.

Fix the configure test 'supports setting throwSuggestions' to assert
using the new error message produced.

BREAKING CHANGE: The errors thrown by getBy queries will have slightly
different messages as they are no longer being processed directly by
WebdriverIO. This may cause tests that assert against the error message
to fail after this fix.