fix: injecting DOM Testing Library on Firefox
Evaluating the UMD script is not enough to add TestingLibraryDom to the
window when running on Firefox. Add a script tag to the head element
with the UMD inlined to add TestingLibraryDom to the window.

Why it doesn't work is not clear to me right now so there may be a
better way of fixing this. I'm assuming this is a security feature of
Firefox but that is a guess.