test: increase timeout for delayed button elements
The support waitFor options test is currently flaky, it could be that
the buttons are appearing too quickly for the test to pass reliably.
Increase the timeout of delayed buttons to 750ms, which is 250ms longer
than currently, but still below the default timeout that an async util
will wait of 1s.