build: fix eslint errors after kcd-scripts update
Most eslint warnings are related to the
@typescript-eslint/no-explicit-any rule being switched on. It is
possible to replace most uses of any with specific types, however there
are two main areas where that is not possible:

- The stub WebdriverIO types are intentionally loose to support multiple
versions of WebdriverIO as well as differences between sync and async
mode types.

- The setupBrowser function does not currently have any way to know
which query it is setting up in each iteration of the forEach, resulting
in issues when typing the resulting query. This can't be resolved
without a refactor of how the queries are set up. Likely each variant of
query will need their own version of createQuery, e.g.
createFindAllByQuery, createFindByQuery etc.

There are additional errors resulting from the parserOptions.project
property being set in the base eslint.js file. It looks like kcd-scripts
assumes that there is a single tsconfig, whereas we need multiple
tsconfigs to support the different WebdriverIO modes in our async and
sync tests. This is fixed by overriding the parserOptions.project
option to be a glob that matches all our tsconfigs.