fix: typescript compilation when checking libs
When the skipLibCheck Typescript compiler option is set to false
compilation fails. This is due to the global WebdriverIO.Element
interface being extended with ElementBase: the compiler complains that
the types of $, execute and executeAsync are not identical to those
found in the webdriverio package.

To fix the compilation the WebdriverIO.Element interface either needs to
be extended with identical types as those used in the webdriverio
package or we stop extending it with ElementBase entirely.

The first option is not realistic as at build time we don't know which
version of the webdriverio package the end user will have installed. We
can however stop extendng the interface: WebdriverIO.Element is only
extended with ElementBase to allow an element to be passed to the within
method internally, if instead we coerce the WebdriverIO.Element to be an
ElementBase we no longer need to extend the interface.

This does not cause any loss in type safety either as we were
effectively coercing the type to be the same as ElementBase when we
extended the WebdriverIO.Element interface anyways.

Change the test tsconfigs to have skipLibCheck set to false so that the
typecheck script catches these types of bugs in future.