fix: injecting DTL when amd define exists
https://www.ebay.com has a global define function with define.amd as an
empty object. This causes the UMD distribution of Dom Testing Library to
use the define function to expose it's methods. Since we assume the
methods are available on the TestingLibraryDom object this breaks the
library.

Modify the UMD distribution so it creates the TestingLibraryDom even
when a define function exists.