build: fix chromedriver version in CI
GitHub Actions VMs have Chrome installed, but there is a period of time
after a new version is published that the machine is behind and builds
fail.

Specify the version used by GitHub Actions in our wdio.conf.js when CI
is true until a better solution can be used.

Ideally we'd either detect the correct version when running in CI or
install our own version of chrome-stable. These options are more
desirable than setting the Chrome version because we want to be aware of
bugs that come up with newer browser versions.