test: fix element refetch test using Puppeteer
When using Puppeteer to drive the browser it adds extra attributes to
identify the element. This breaks the test for refetching elements as we
are checking the stringified elements are the same, however the
element identifier added by Puppeteer has changed.

Check that the refetched element has the correct text instead.