test: add failing test cases for prod React apps
When testing a production React application Puppeteer is unable to
serialise the elements due to circular references from React Fiber.

The tests include a basic `getByText` test and another for checking
refetching after a key change. The later is important as key changes are
common in React.

Skip the tests until they pass with Puppeteer as well.