build: prevent chrome version mismatch in actions
When using the default selenium-standalone Chromedriver version there is
a period of time where the Ubuntu container's Chrome version is behind
by a major version, this causes the tests to fail when using the
selenium-standalone service.

Find the version of Chromedriver that is on the container and then use
that to specify the Chromedriver version to use in the tests.