fix: null queries with prod React and Puppeteer (#38)
When returning queried elements directly from executeAsync Puppeteer
will fail and return null if the element contains circular references. Elements
in production React apps contain circular references due to React Fiber.
Puppeteer is the default driver used by wdio when no services are defined.

Instead of returning elements directly when Simmer fails add a data
attribute with an id to use as a selector. Simmer should remain the primary
method of generating selectors as it is more tolerant to scenarios where
the element is regenerated, such as when React finds a key change.

Co-authored-by: olivierwilkinson <olivier.wilkinson@gmail.com>