build: avoid failing actions from old chromedriver
When the GitHub actions chrome version is updated there is not an easy
way to update the version of Chromedriver in the package.json to match.

Remove chromedriver from dev dependencies and instead use the version
that is installed by selenium-standalone when running chromedriver tests
locally.

Ensure the correct version is installed in CI by specifying which
version of chromedriver to install in the prepare script.