fix: error running with wdio v8  (#47)
Closes issue #46.

When using WebdriverIO v8 we encounter `TypeError: container.parent.execute is not a function`. This is due to the execute function being moved off of Elements. Traverse up through the Element's parents until we find one with `execute` and use that to inject DTL and to execute the query. Using the parent to execute does not change the behaviour of `within` since the passed Element is still used as the container for the query.